+++ README.md
@@ -1,124 +1,36 @@
-# Next Phisher
+This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
 
-# Table of Contents
-- [Description](#Description)
-- [Disclaimer](#Disclaimer)
-- [Features](#Features)
-  - [Current Features](#Current-Features)  
-- [Cloning](#Cloning)
-- [Setup](#Setup)
-- [Start Up](#Start-Up)
-- [Usage](#Usage)
+## Getting Started
 
-# Description
-This project is a Phishing Detection & Alert System, currently comprising a phishing website and a decoy website. The phishing site collects data locally and offers an admin dashboard for monitoring. It also generates QR codes linking to both sites. The decoy site allows custom alerts to be sent to users. Installation and setup are streamlined via scripts in the package JSON. Additionally, WebSocket integration enables real-time communication for alerts and future data transmission. This system aims to evolve into a comprehensive cybersecurity solution.
+First, run the development server:
 
-# Disclaimer
-This project is intended solely for educational purposes and ethical research into cybersecurity. Any misuse, including but not limited to, deploying this system for malicious activities, is strictly prohibited and may result in legal consequences. Please be aware that various measures, including breadcrumbs, have been intentionally integrated into the project to ensure that any unauthorized or malicious use can be readily detected. Engaging in unauthorized activities with this project could lead to identification and prosecution. We strongly advise users to utilize this project responsibly and in accordance with applicable laws and regulations. By accessing or using this project, you acknowledge and agree to abide by these terms and to use the system only for lawful and ethical purposes.
-
-# Features
-
-## Current Features
-1. notGoogle Phishing Site:
-- A simulated phishing site named “notGoogle” serves as a decoy to prevent users from inadvertently inputting sensitive data.
-- Enhancement: Consider adding educational content to raise awareness about phishing risks and provide clear warnings.
-
-2. Decoy Site with Custom Alerts:
-- The decoy site allows administrators to send custom alerts from the decoy controller screen in the admin dashboard.
-
-3. Local Storage for Collected User Data:
-- User data collected during phishing input is stored locally.
-
-4. QR Codes for Service Links:
-- QR codes are provided, linking to each service in the Admin Dashboard.
-
-5. User Redirection Abilities:
-- Allow users to be redirected to different sites while on the decoy site and subsequently on the phishing sites.
-
-6. Alerts for Phishing Websites:
-- Enable sending alerts to the phishing websites themselves.
-
-# Cloning
-1. Clone repository(git ssh method) run this command: 
 ```bash
-git clone git@github.com:Pmacdon15/nextphisher.git
+npm run dev
+# or
+yarn dev
+# or
+pnpm dev
+# or
+bun dev
 ```
 
-# Setup
+Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
-After cloning run: 
-```Bash
-cd nextpisher
-```
+You can start editing the page by modifying `app/page.js`. The page auto-updates as you edit the file.
 
-Once you are inside of the project directory you will have to run a few commands from the root directory to set the project up.
-The commands are:
+This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
 
-1. Install dependencies from both client and server in one command:
- ```Bash
- npm run install
- ```
-> **Note**
-> The same as running npm i in both the client and server.
+## Learn More
 
-2. Setup .env files in both client and server:
- ```Bash
- npm run setup
- ```
- It will run two scripts and ask you a series of questions:
-- JWT Secret key can be anything thing, it is used for encoding and decoding authentication tokens.
-- Admin password for logging in to the Admin Dashboard(you will need to later to use the tools).
-- Back end port number(you will need this later for setting up port forwarding).
-- Back end Ip Address(The Ip address you are running the server on).
-> **Note**
-> Use the same Backend port number when asked during the scripts execution.
+To learn more about Next.js, take a look at the following resources:
 
-3. Now we have to build the client easily done by running:
-```Bash
-npm run build-client
-```
+- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
+- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
 
-4. Setup port forwarding on your router for both port 3000 and the answer you gave during setup.
+You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
 
-# Start up
+## Deploy on Vercel
 
-Front end easily started from the root by running:
-```Bash
-npm run start-client
-```
+The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
 
-Server started by:
-```Bash
-npm run start-server
-```
-
-Using pm2 works well is use these commands to start the client and server:
-```Bash
-pm2 start npm --name nextPisher -- run start-client
-pm2 start npm --name nextPisherServer -- run start-server
-```
-
-# Usage
-
-## Links
-
-### Client Side
-- notGoogle(phishing) User data will be stored when entered on this page.
-```HTML
-http://"Your backend Ip here":3000/notGoogle
-```
-(Remove "") 
-
-- decoy(User interaction) Admin can send custom alerts to clients on this site.
-```HTML
-http://"Your backend Ip here":3000/decoy
-```
-(Remove "")
-
-### Admin Side
-- Admin Dashboard
-```HTML
-http://"Your backend Ip here":3000/
-```
-(Remove "")
-You will be asked to Login with the password you set earlier. Through the Admin Dashboard you will be able to access a number of features such as getting the Qr code for active services allowing users to link to the sites from the Admin Dashboard for ease of use. A link to the stored user data page, collected from the phishing pages. Finally the decoy controller page which allows you to send custom alerts to clients on the decoy page.
+Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
