--- client/package.json
@@ -1,30 +1,0 @@
-{
-  "type": "module",
-  "name": "nextphisher",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "node server.js",
-    "build": "next build",
-    "start": "set NODE_ENV=production&& node server.js",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/material": "^5.15.15",
-    "dotenv": "^16.4.5",
-    "jsonwebtoken": "^9.0.2",
-    "next": "14.1.4",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-hook-form": "^7.51.2",
-    "sharp": "^0.33.3",
-    "socket.io": "^4.7.5",
-    "socket.io-client": "^4.7.5"
-  },
-  "devDependencies": {
-    "eslint": "^8",
-    "eslint-config-next": "14.1.4"
-  }
-}
