--- server/package.json
@@ -1,13 +1,0 @@
-{
-  "dependencies": {
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "jsonwebtoken": "^9.0.2",
-    "nodemon": "^3.1.0",
-    "socket.io": "^4.7.5",
-    "socket.io-client": "^4.7.5"
-  },
-  "scripts": {
-    "start": "node server.js"
-  }
-}
