--- server/setup.js
@@ -1,32 +1,0 @@
-const fs = require('fs');
-const readline = require('readline');
-
-const rl = readline.createInterface({
-  input: process.stdin,
-  output: process.stdout
-});
-
-const questions = [  
-  'BACK_END_PORT'
-];
-
-function askQuestions(index, answers) {
-  if (index >= questions.length) {
-    writeToFile(answers);
-    rl.close();
-    return;
-  }
-
-  rl.question(`${questions[index]} = `, (answer) => {
-    answers.push(`${questions[index]} = ${answer}`);
-    askQuestions(index + 1, answers);
-  });
-}
-
-function writeToFile(answers) {
-  const content = answers.join('\n');
-  fs.writeFileSync('.env', content); // Write to .env file
-  console.log('Config file created successfully!');
-}
-
-askQuestions(0, []);
