+++ src/app/admin/QrCode/[projectName]/page.jsx
