+++ src/app/admin/decoy/page.jsx
@@ -52,24 +52,12 @@ const Home = () => {
       userList = userList.filter((user) => user !== "Decoy Controller");
       setUserList(userList);
     }
-    // function handleAlert(message){
-    //   console.log("Alerting user with message: ", message);
-    //   alert(message);
-    // }
-
-    // function handlePushToPage(site) {      
-    //   router.push(site);
-    // }
 
     socket.on("connect", onConnect);
-    // socket.on("alert", handleAlert);
-    // socket.on("pushToPage", handlePushToPage);
     socket.on("disconnect", onDisconnect);
 
     return () => {
       socket.off("connect", onConnect);
-      // socket.off("alert", handleAlert);      
-      // socket.off("pushToPage", handlePushToPage);
       socket.off("UserList", handleUserList);
       socket.off("disconnect", onDisconnect);
     };
@@ -107,53 +95,6 @@ const Home = () => {
     }
   }, []);
 
-  // const [isConnected, setIsConnected] = useState(false);
-  // const [transport, setTransport] = useState("N/A");
-  // const [userList, setUserList] = useState([]);
-
-  // useEffect(() => {
-  //   if (socket.connected) {
-  //     onConnect();
-  //   }
-
-  //   function onConnect() {
-  //     setIsConnected(true);
-  //     setTransport(socket.io.engine.transport.name);
-
-  //     socket.io.engine.on("upgrade", (transport) => {
-  //       setTransport(transport.name);
-  //     });
-
-  //     socket.emit("join", "Decoy");
-  //     socket.on("UserList", handleUserList);
-  //   }
-
-  //   function onDisconnect() {
-  //     setIsConnected(false);
-  //     setTransport("N/A");
-  //     // Remove event listener when disconnecting
-  //     socket.off("UserList", handleUserList);
-  //   }
-
-  //   // Define function to handle "UserList" event
-  //   function handleUserList(userList) {
-  //     userList = userList.filter(
-  //       (user, index) => userList.indexOf(user) === index && user !== null
-  //     );
-  //     userList = userList.filter((user) => user !== "Decoy Controller");
-  //     setUserList(userList);
-  //   }
-   
-  //   socket.on("connect", onConnect);
-  //   socket.on("disconnect", onDisconnect);
-
-  //   return () => {
-  //     socket.off("connect", onConnect);
-  //     socket.off("disconnect", onDisconnect);
-  //     socket.off("UserList", handleUserList);
-  //   };
-  // }, []);
-
   const handleAlertClick = (userId) => {
     if (socket) {
       console.log("Alerting user with message: ", message);
@@ -180,6 +121,7 @@ const Home = () => {
       console.error("Socket is not initialized.");
     }
   };
+
 console.log(currentUser)
   return (
     <div className={adminStyles.container}>
